requirements.txt
pillow
ttkthemes